Bootstrap Confidence Intervals

This function will conduct a bootstrap routine for each predictor or beta value for the logistic regression curve. The user is able to choose the amount of bootstraps to conduct as well as the confidence level for the intervals.

Boot_CI(n = 20, alpha = 0.05, pred, resp)

Arguments

n

A value(integer) which will be the number of bootstraps run in the function. The default value is 20.

alpha

A value(numeric) that can be between 0 and 1, but is usually closer to 0. The default value is 0.05.

pred

A matrix(numeric) containing variables that will be used to predict the response.

resp

A vector(numeric) containing values that represent a response variable.

Value

A matrix(numeric) where the number of rows is equal to the amount of betas and 2 columns. Each row represents a confidence interval for each predictor at the chosen confidence level.

Author

Yizhong Zhang, Alexander Miller

Examples

mydata <- read.csv("expenses.csv")
#> Warning: cannot open file 'expenses.csv': No such file or directory
#> Error in file(file, "rt"): cannot open the connection
mydata$sex <- factor(mydata$sex, levels = c("male","female"), labels = c(0,1))
#> Error in eval(expr, envir, enclos): object 'mydata' not found
mydata$smoker <- factor(mydata$smoker, levels = c("yes","no"), labels = c(1,0))
#> Error in eval(expr, envir, enclos): object 'mydata' not found
mydata$region_northeast <- ifelse(mydata$region=="northeast",1,0)
#> Error in eval(expr, envir, enclos): object 'mydata' not found
mydata$region_northwest <- ifelse(mydata$region=="northwest",1,0)
#> Error in eval(expr, envir, enclos): object 'mydata' not found
mydata$region_southeast <- ifelse(mydata$region=="southeast",1,0)
#> Error in eval(expr, envir, enclos): object 'mydata' not found
mydata$region_northeast <- ifelse(mydata$region=="northeast",1,0)
#> Error in eval(expr, envir, enclos): object 'mydata' not found
mydata$region_northwest <- ifelse(mydata$region=="northwest",1,0)
#> Error in eval(expr, envir, enclos): object 'mydata' not found
mydata$region_southeast <- ifelse(mydata$region=="southeast",1,0)
#> Error in eval(expr, envir, enclos): object 'mydata' not found
mydata <- mydata[,-6]
#> Error in eval(expr, envir, enclos): object 'mydata' not found
mydata <- as.matrix(mydata)
#> Error in eval(expr, envir, enclos): object 'mydata' not found
mydata <- as.numeric(mydata)
#> Error in eval(expr, envir, enclos): object 'mydata' not found
mydata <- matrix(mydata, ncol=9)
#> Error in eval(expr, envir, enclos): object 'mydata' not found
X <- mydata[, c(1:4,6:9)]
#> Error in eval(expr, envir, enclos): object 'mydata' not found
Y <- mydata[, 5]
#> Error in eval(expr, envir, enclos): object 'mydata' not found
Boot_CI_demo <- Boot_CI(n=20, alpha=0.05, pred= X, resp= Y)
#> Error in eval(expr, envir, enclos): object 'X' not found