Metric Plot

This function conducts logistic regression on data and plots a given metric that visualizes the performance of the regression.

plot_metric(resp, pred, metric)

Arguments

resp

A vector(numeric) containing values that represent a response variable.

pred

A matrix(numeric) containing variables that will be used to predict the response.

metric

This can be any of the following metrics: "Prevalence", "Accuracy","Sensitivity", "Specificity", "False Discovery Rate","Diagnostic Odds Ratio"

Value

A plot that shows the entered metric over a grid of cut-off values for prediction going from 0.1 to 0.9 with steps of 0.1.

Author

Yizhong Zhang, Alexander Miller

Examples

mydata <- read.csv("expenses.csv")
#> Warning: cannot open file 'expenses.csv': No such file or directory
#> Error in file(file, "rt"): cannot open the connection
mydata$sex <- factor(mydata$sex, levels = c("male","female"), labels = c(0,1))
#> Error in eval(expr, envir, enclos): object 'mydata' not found
mydata$smoker <- factor(mydata$smoker, levels = c("yes","no"), labels = c(1,0))
#> Error in eval(expr, envir, enclos): object 'mydata' not found
mydata$region_northeast <- ifelse(mydata$region=="northeast",1,0)
#> Error in eval(expr, envir, enclos): object 'mydata' not found
mydata$region_northwest <- ifelse(mydata$region=="northwest",1,0)
#> Error in eval(expr, envir, enclos): object 'mydata' not found
mydata$region_southeast <- ifelse(mydata$region=="southeast",1,0)
#> Error in eval(expr, envir, enclos): object 'mydata' not found
mydata$region_northeast <- ifelse(mydata$region=="northeast",1,0)
#> Error in eval(expr, envir, enclos): object 'mydata' not found
mydata$region_northwest <- ifelse(mydata$region=="northwest",1,0)
#> Error in eval(expr, envir, enclos): object 'mydata' not found
mydata$region_southeast <- ifelse(mydata$region=="southeast",1,0)
#> Error in eval(expr, envir, enclos): object 'mydata' not found
mydata <- mydata[,-6]
#> Error in eval(expr, envir, enclos): object 'mydata' not found
mydata <- as.matrix(mydata)
#> Error in eval(expr, envir, enclos): object 'mydata' not found
mydata <- as.numeric(mydata)
#> Error in eval(expr, envir, enclos): object 'mydata' not found
mydata <- matrix(mydata, ncol=9)
#> Error in eval(expr, envir, enclos): object 'mydata' not found
X <- mydata[, c(1:4,6:9)]
#> Error in eval(expr, envir, enclos): object 'mydata' not found
Y <- mydata[, 5]
#> Error in eval(expr, envir, enclos): object 'mydata' not found
plot_metric(resp=Y, pred=X, metric = "Prevalence")
#> Error in eval(expr, envir, enclos): object 'X' not found
plot_metric(resp=Y, pred=X, metric = "Accuracy")
#> Error in eval(expr, envir, enclos): object 'X' not found